requirements.txt
python-dotenv
# alith-sdk 
# (Conceptual dependency, commented out since it's being mocked)